- added real-time monitoring activity for viewing C/N0 and AGC (Service interactions can be slightly buggy at start-up and shutdown, but the bugs should be hidden right now by excessive exception catching)

+++ torgi/src/main/java/org/sofwerx/torgi/service/GeoPackageRecorder.java
@@ -271,7 +271,7 @@ public class GeoPackageRecorder extends HandlerThread {
 
                         satrow.setValue("svid", g.getSvid());
                         satrow.setValue("constellation", con);
-                        satrow.setValue("cn0", (double) g.getCn0DbHz());
+                        satrow.setValue("cn0", g.getCn0DbHz());
 
                         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
                             if (g.hasAutomaticGainControlLevelDb()) {
